fix: split i/p/u/fs/squashfs on buildcfg dep
